Refactor storage hooks to be DRY
