+++ Client/src/Hooks/SessionBackedState.ts
@@ -1,38 +1,42 @@
-import { useCallback, useState } from 'react';
+// import { useCallback, useState } from 'react';
+import { useStorageBackedState } from './StorageBackedState';
+import { partial } from 'lodash';
 
-type Encoder<T> = (t: T) => string;
-type Parser<T> = (s: string) => T;
+export const useSessionBackedState = partial(useStorageBackedState, window.sessionStorage)
 
-export function useSessionBackedState<T>(
-  defaultValue: T,
-  key: string,
-  encode: Encoder<T>,
-  parse: Parser<T>,
-): [ T, (newState: T) => void ] {
-  let initialValue = defaultValue;
+// type Encoder<T> = (t: T) => string;
+// type Parser<T> = (s: string) => T;
 
-  try {
-    const storedStringValue = window.sessionStorage.getItem(key);
+// export function useSessionBackedState<T>(
+//   defaultValue: T,
+//   key: string,
+//   encode: Encoder<T>,
+//   parse: Parser<T>,
+// ): [ T, (newState: T) => void ] {
+//   let initialValue = defaultValue;
 
-    if (storedStringValue !== null) {
-      initialValue = parse(storedStringValue);
-    }
-  } catch (e) {
-    console.warn(
-      `Failed attempt to retrieve state value at session key ${key}: ${e}; falling back to component state`
-    );
-  }
+//   try {
+//     const storedStringValue = window.sessionStorage.getItem(key);
 
-  const [ state, setState ] = useState(initialValue);
+//     if (storedStringValue !== null) {
+//       initialValue = parse(storedStringValue);
+//     }
+//   } catch (e) {
+//     console.warn(
+//       `Failed attempt to retrieve state value at session key ${key}: ${e}; falling back to component state`
+//     );
+//   }
 
-  const setSessionBackedState = useCallback((newValue: T) => {
-    try {
-      window.sessionStorage.setItem(key, encode(newValue));
-    } catch {
-      console.warn(`Failed attempt to persist state value ${newValue} at session key ${key}; falling back to component state`);
-    }
-    setState(newValue);
-  }, [ encode ]);
+//   const [ state, setState ] = useState(initialValue);
 
-  return [ state, setSessionBackedState ];
-};
+//   const setSessionBackedState = useCallback((newValue: T) => {
+//     try {
+//       window.sessionStorage.setItem(key, encode(newValue));
+//     } catch {
+//       console.warn(`Failed attempt to persist state value ${newValue} at session key ${key}; falling back to component state`);
+//     }
+//     setState(newValue);
+//   }, [ encode ]);
+
+//   return [ state, setSessionBackedState ];
+// };
